// backend/src/models/schemas/User.schema.ts
import { ObjectId } from 'mongodb'
import { UserVerificationStatus } from '~/constants/enums'

export type UserType = {
  _id?: ObjectId
  fullname: string
  email: string
  phone: string
  date_of_birth: Date
  password: string
  created_at: Date
  updated_at: Date
  email_verify_token: string
  forgot_password_token: string
  verify: UserVerificationStatus
  bio: string
  location: string
  website: string
  username: string
  avatar: string
  cover_photo: string
  role: 'admin' | 'member' | 'guest'
}

export default class User {
  _id?: ObjectId
  fullname: string
  email: string
  phone: string
  date_of_birth: Date
  password: string
  created_at: Date
  updated_at: Date
  email_verify_token: string
  forgot_password_token: string
  verify: UserVerificationStatus
  bio: string
  location: string
  website: string
  username: string
  avatar: string
  cover_photo: string
  role: 'admin' | 'member' | 'guest'

  constructor(user: UserType) {
    const date = new Date()
    this._id = user._id || new ObjectId()
    this.fullname = user.fullname || ''
    this.email = user.email
    this.phone = user.phone || ''
    this.date_of_birth = user.date_of_birth || new Date()
    this.password = user.password
    this.created_at = user.created_at || date
    this.updated_at = user.updated_at || date
    this.email_verify_token = user.email_verify_token || ''
    this.forgot_password_token = user.forgot_password_token || ''
    this.verify = user.verify || UserVerificationStatus.Unverified
    this.bio = user.bio || ''
    this.location = user.location || ''
    this.website = user.website || ''
    this.username = user.username || ''
    this.avatar = user.avatar || ''
    this.cover_photo = user.cover_photo || ''
    this.role = user.role || 'guest'
  }
}